+++ Example/CollectionViewController.swift
@@ -55,3 +55,19 @@ class CollectionViewController: UICollectionViewController {
         ]
     }
 }
+
+// MARK: View Model Provider
+
+extension CollectionViewController {
+    /// Pure function mapping new state to a new `FluxCollectionViewModel`.  This is invoked each time the state updates
+    /// in order for ReactiveLists to update the UI.
+    static func viewModel(forState groups: [UserGroup], onDeleteClosure: @escaping (User) -> Void) -> FluxCollectionViewModel {
+        let sections: [FluxCollectionViewModel.SectionModel] = groups.map { group in
+            let cellViewModels = group.users.map { CollectionUserCell(user: $0, onDeleteClosure: onDeleteClosure) }
+            return FluxCollectionViewModel.SectionModel(cellViewModels: cellViewModels, headerHeight: nil, footerHeight: nil)
+        }
+        return FluxCollectionViewModel(sectionModels: sections)
+    }
+}
+
+
